+++ poem/src/listener/acme/listener.rs
@@ -177,7 +177,7 @@ async fn issue_cert(
     // trigger challenge
     let mut valid = false;
 
-    for _ in 0..10 {
+    for i in 1..5 {
         let mut all_valid = true;
 
         for auth_url in &order_resp.authorizations {
@@ -243,7 +243,7 @@ async fn issue_cert(
             break;
         }
 
-        tokio::time::sleep(Duration::from_secs(5)).await;
+        tokio::time::sleep(Duration::from_secs(i * 10)).await;
     }
 
     if !valid {